+++ hibernate-core/src/main/java/org/hibernate/query/sqm/mutation/spi/AfterUseAction.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.query.sqm.mutation.internal.temptable;
+package org.hibernate.query.sqm.mutation.spi;
 
 /**
  * Actions to perform in regard to a temporary table after each use.
