+++ hibernate-core/src/main/java/org/hibernate/query/sqm/mutation/spi/BeforeUseAction.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.query.sqm.mutation.internal.temptable;
+package org.hibernate.query.sqm.mutation.spi;
 
 /**
  * Actions to perform in regards to a temporary table prior to each use.
